+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Sheep
 
 - improve handling of kernel module - this prevents an endless reload-loop
 - prevent early exit if power-tracing is disabled (10 s missing)
+- ~~transform pru_util into a monitor-thread to prevent overfilling buffer during waiting~~
+  - bad idea as a non-atomic monitor-thread corrupts shared-memory-access
+  - better: add pre-experiment-loop that collects pru-util data
 
 Ansible